+++ packages/link/__tests__/web/link.web.test.js
@@ -1,57 +1,37 @@
 import React from "react";
-import TestRenderer from "react-test-renderer";
+import { fireEvent, render } from "@testing-library/react";
 import { colours, fonts, fontSizes } from "@times-components/styleguide";
-import { iterator } from "@times-components/test-utils";
+import "@testing-library/jest-dom";
 import Link from "../../src/link";
-import shared from "../shared";
 
-require("jest-styled-components");
-
-const tests = [
-  {
-    name: "with responsive styles",
-    test() {
-      const responsiveLinkStyles = {
-        base: `
-        color: ${colours.functional.action};
-        font-family: "${fonts.bodyRegular}";
-        font-size: ${fontSizes.bodyMobile}px;
-        line-height: 26px;
-        margin-bottom: 25px;
-        margin-top: 0;
-    `,
-        medium: `
-        font-size: ${fontSizes.body}px;
-        line-height: 30px;
-    `
-      };
-
-      const testInstance = TestRenderer.create(
-        <Link
-          onPress={() => {}}
-          responsiveLinkStyles={responsiveLinkStyles}
-          url="http://thetimes.co.uk"
-        >
-          The Times
-        </Link>
-      );
-
-      expect(testInstance).toMatchSnapshot();
-    }
-  },
-  {
-    name: "with a target",
-    test() {
-      const tree = TestRenderer.create(
-        <Link onPress={() => {}} target="_blank" url="http://thetimes.co.uk">
-          The Times
-        </Link>
-      );
-
-      expect(tree).toMatchSnapshot();
-    }
-  },
-  ...shared("p")
-];
-
-iterator(tests);
+describe("Link", () => {
+  it("renders the link", () => {
+    const { baseElement, getByText, getByRole } = render(
+      <Link onPress={() => {}} url="http://thetimes.co.uk">
+        The Times
+      </Link>
+    );
+    expect(baseElement).toMatchSnapshot();
+    expect(getByText('The Times')).toBeVisible();
+    expect(getByRole("link")).not.toHaveAttribute("target", "_blank");
+  });
+  it("renders the link with a target", () => {
+    const { baseElement, getByRole } = render(
+      <Link onPress={() => {}} target="_blank" url="http://thetimes.co.uk">
+        The Times
+      </Link>
+    );
+    expect(baseElement).toMatchSnapshot();
+    expect(getByRole("link")).toHaveAttribute("target", "_blank");
+  });
+  it("calls onPress prop when clicked", () => {
+    const mockOnPress = jest.fn();
+    const { getByRole } = render(
+      <Link onPress={mockOnPress} target="_blank" url="http://thetimes.co.uk">
+        The Times
+      </Link>
+    );
+    fireEvent.click(getByRole("link"));
+    expect(mockOnPress).toHaveBeenCalled();
+  });
+});
